+++ LukyNow/Assets/ScratchCards/Scripts/ScratchReward.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ScratchReward : Reward
 {
-    public ScratchReward(int dollars , double coin)
+    public ScratchReward(double dollars , double coin)
     {
         _rewardDollars = dollars;
         _rewardCoin = coin;
